Moved some code in the module risklib.asset [skip hazardlib]
