+++ openquake/risklib/riskinput.py
@@ -25,7 +25,7 @@ from openquake.baselib import hdf5
 from openquake.baselib.python3compat import zip, encode, decode
 from openquake.baselib.general import groupby, get_array, AccumDict
 from openquake.hazardlib import site, calc, valid
-from openquake.risklib import scientific, riskmodels
+from openquake.risklib import scientific, riskmodels, asset
 
 
 class ValidationError(Exception):
@@ -70,240 +70,6 @@ def get_refs(assets, hdf5path):
     """
     with hdf5.File(hdf5path, 'r') as f:
         return f['asset_refs'][[a.idx for a in assets]]
-
-
-class AssetCollection(object):
-    D, I, R = len('deductible-'), len('insurance_limit-'), len('retrofitted-')
-
-    def __init__(self, assets_by_site, assets_by_tag, cost_calculator,
-                 time_event, time_events=''):
-        self.cc = cost_calculator
-        self.time_event = time_event
-        self.time_events = time_events
-        self.tot_sites = len(assets_by_site)
-        self.array = self.build_asset_collection(assets_by_site, time_event)
-        dic = dict(zip(self.array['idx'], range(len(self.array))))
-        self.tagnames = assets_by_tag.tagnames
-        self.aids_by_tag = {}
-        for tag, idxs in assets_by_tag.items():
-            aids = []
-            for idx in idxs:
-                try:
-                    aids.append(dic[idx])
-                except KeyError:  # discarded by assoc_assets_sites
-                    continue
-            self.aids_by_tag[tag] = U32(aids)
-        fields = self.array.dtype.names
-        self.loss_types = [f[6:] for f in fields if f.startswith('value-')]
-        if 'occupants' in fields:
-            self.loss_types.append('occupants')
-        self.loss_types.sort()
-        self.deduc = [n for n in fields if n.startswith('deductible-')]
-        self.i_lim = [n for n in fields if n.startswith('insurance_limit-')]
-        self.retro = [n for n in fields if n.startswith('retrofitted-')]
-
-    @property
-    def taxonomies(self):
-        """
-        Return a list of taxonomies, one per asset (with duplicates)
-        """
-        if not hasattr(self, '_taxonomy'):
-            self._taxonomy = [None] * len(self)
-            for tag, aids in self.aids_by_tag.items():
-                name, value = tag.split('=', 1)
-                if name == 'taxonomy':
-                    for aid in aids:
-                        self._taxonomy[aid] = value
-        return self._taxonomy
-
-    def tags(self):
-        """
-        :returns: list of sorted tags
-        """
-        return sorted(self.aids_by_tag)
-
-    def units(self, loss_types):
-        """
-        :param: a list of loss types
-        :returns: an array of units as byte strings, suitable for HDF5
-        """
-        units = self.cc.units
-        lst = []
-        for lt in loss_types:
-            if lt.endswith('_ins'):
-                lt = lt[:-4]
-            lst.append(encode(units[lt]))
-        return numpy.array(lst)
-
-    def assets_by_site(self):
-        """
-        :returns: numpy array of lists with the assets by each site
-        """
-        assets_by_site = [[] for sid in range(self.tot_sites)]
-        for i, ass in enumerate(self.array):
-            assets_by_site[ass['site_id']].append(self[i])
-        return numpy.array(assets_by_site)
-
-    def values(self, aids=None):
-        """
-        :param aids: asset indices where to compute the values (None means all)
-        :returns: a structured array of asset values by loss type
-        """
-        if aids is None:
-            aids = range(len(self))
-        loss_dt = numpy.dtype([(str(lt), F32) for lt in self.loss_types])
-        vals = numpy.zeros(len(aids), loss_dt)  # asset values by loss_type
-        for i, aid in enumerate(aids):
-            asset = self[aid]
-            for lt in self.loss_types:
-                vals[i][lt] = asset.value(lt, self.time_event)
-        return vals
-
-    def tagmask(self):
-        """
-        :returns: array of booleans of shape (A, T)
-        """
-        tags = self.tags()
-        tagidx = {t: i for i, t in enumerate(tags)}
-        mask = numpy.zeros((len(self), len(tags)), bool)
-        for tag, aids in self.aids_by_tag.items():
-            mask[aids, tagidx[tag]] = True
-        return mask
-
-    def get_tax_idx(self):
-        """
-        :returns: list of tag indices corresponding to taxonomies
-        """
-        return [i for i, t in enumerate(self.tags())
-                if t.startswith('taxonomy=')]
-
-    def __iter__(self):
-        for i in range(len(self)):
-            yield self[i]
-
-    def __getitem__(self, aid):
-        a = self.array[aid]
-        values = {lt: a['value-' + lt] for lt in self.loss_types
-                  if lt != 'occupants'}
-        if 'occupants' in self.array.dtype.names:
-            values['occupants_' + str(self.time_event)] = a['occupants']
-        return riskmodels.Asset(
-                a['idx'],
-                self.taxonomies[aid],
-                number=a['number'],
-                location=(valid.longitude(a['lon']),  # round coordinates
-                          valid.latitude(a['lat'])),
-                values=values,
-                area=a['area'],
-                deductibles={lt[self.D:]: a[lt] for lt in self.deduc},
-                insurance_limits={lt[self.I:]: a[lt] for lt in self.i_lim},
-                retrofitteds={lt[self.R:]: a[lt] for lt in self.retro},
-                calc=self.cc, ordinal=aid)
-
-    def __len__(self):
-        return len(self.array)
-
-    def __toh5__(self):
-        # NB: the loss types do not contain spaces, so we can store them
-        # together as a single space-separated string
-        attrs = {'time_event': self.time_event or 'None',
-                 'time_events': ' '.join(map(decode, self.time_events)),
-                 'loss_types': ' '.join(self.loss_types),
-                 'deduc': ' '.join(self.deduc),
-                 'i_lim': ' '.join(self.i_lim),
-                 'retro': ' '.join(self.retro),
-                 'tot_sites': self.tot_sites,
-                 'tagnames': encode(self.tagnames),
-                 'nbytes': self.array.nbytes}
-        tags, all_aids = [], []
-        for tag, aids in sorted(self.aids_by_tag.items()):
-            tags.append(tag)
-            all_aids.append((aids,))
-        return dict(array=self.array,
-                    tags=numpy.array(tags, hdf5.vstr),
-                    aids=numpy.array(all_aids, aids_dt),
-                    cost_calculator=self.cc), attrs
-
-    def __fromh5__(self, dic, attrs):
-        for name in ('time_events', 'loss_types', 'deduc', 'i_lim', 'retro'):
-            setattr(self, name, attrs[name].split())
-        self.tagnames = attrs['tagnames']
-        self.time_event = attrs['time_event']
-        self.tot_sites = attrs['tot_sites']
-        self.nbytes = attrs['nbytes']
-        self.array = dic['array'].value
-        self.cc = dic['cost_calculator']
-        # dic['aids'] is an array of dtype `aids_dt` with field 'aids'
-        self.aids_by_tag = dict(zip(dic['tags'], dic['aids']['aids']))
-
-    @staticmethod
-    def build_asset_collection(assets_by_site, time_event=None):
-        """
-        :param assets_by_site: a list of lists of assets
-        :param time_event: a time event string (or None)
-        :returns: an array `assetcol`
-        """
-        for assets in assets_by_site:
-            if len(assets):
-                first_asset = assets[0]
-                break
-        else:  # no break
-            raise ValueError('There are no assets!')
-        candidate_loss_types = list(first_asset.values)
-        loss_types = []
-        the_occupants = 'occupants_%s' % time_event
-        for candidate in sorted(candidate_loss_types):
-            if candidate.startswith('occupants'):
-                if candidate == the_occupants:
-                    loss_types.append('occupants')
-                # discard occupants for different time periods
-            else:
-                loss_types.append('value-' + candidate)
-        deductible_d = first_asset.deductibles or {}
-        limit_d = first_asset.insurance_limits or {}
-        retrofitting_d = first_asset.retrofitteds or {}
-        deductibles = ['deductible-%s' % name for name in deductible_d]
-        limits = ['insurance_limit-%s' % name for name in limit_d]
-        retrofittings = ['retrofitted-%s' % n for n in retrofitting_d]
-        float_fields = loss_types + deductibles + limits + retrofittings
-        asset_dt = numpy.dtype(
-            [('idx', U32), ('lon', F32), ('lat', F32), ('site_id', U32),
-             ('number', F32), ('area', F32)] + [
-                 (str(name), float) for name in float_fields])
-        num_assets = sum(len(assets) for assets in assets_by_site)
-        assetcol = numpy.zeros(num_assets, asset_dt)
-        asset_ordinal = 0
-        fields = set(asset_dt.fields)
-        for sid, assets_ in enumerate(assets_by_site):
-            for asset in sorted(assets_, key=operator.attrgetter('idx')):
-                asset.ordinal = asset_ordinal
-                record = assetcol[asset_ordinal]
-                asset_ordinal += 1
-                for field in fields:
-                    if field == 'number':
-                        value = asset.number
-                    elif field == 'area':
-                        value = asset.area
-                    elif field == 'idx':
-                        value = asset.idx
-                    elif field == 'site_id':
-                        value = sid
-                    elif field == 'lon':
-                        value = asset.location[0]
-                    elif field == 'lat':
-                        value = asset.location[1]
-                    elif field == 'occupants':
-                        value = asset.values[the_occupants]
-                    else:
-                        try:
-                            name, lt = field.split('-')
-                        except ValueError:  # no - in field
-                            name, lt = 'value', field
-                        # the line below retrieve one of `deductibles`,
-                        # `insured_limits` or `retrofitteds` ("s" suffix)
-                        value = getattr(asset, name + 's')[lt]
-                    record[field] = value
-        return assetcol
 
 
 def read_composite_risk_model(dstore):
