+++ openquake/risklib/riskmodels.py
@@ -21,212 +21,13 @@ import inspect
 import functools
 import numpy
 
-from openquake.baselib import hdf5
 from openquake.baselib.general import CallableDict
 from openquake.hazardlib import valid
-from openquake.risklib import utils, scientific
+from openquake.risklib import utils, scientific, asset
 
 U32 = numpy.uint32
 F32 = numpy.float32
 registry = CallableDict()
-
-
-class CostCalculator(object):
-    """
-    Return the value of an asset for the given loss type depending
-    on the cost types declared in the exposure, as follows:
-
-        case 1: cost type: aggregated:
-            cost = economic value
-        case 2: cost type: per asset:
-            cost * number (of assets) = economic value
-        case 3: cost type: per area and area type: aggregated:
-            cost * area = economic value
-        case 4: cost type: per area and area type: per asset:
-            cost * area * number = economic value
-
-    The same "formula" applies to retrofitting cost.
-    """
-    def __init__(self, cost_types, area_types, units,
-                 deduct_abs=True, limit_abs=True):
-        if set(cost_types) != set(area_types):
-            raise ValueError('cost_types has keys %s, area_types has keys %s'
-                             % (sorted(cost_types), sorted(area_types)))
-        for ct in cost_types.values():
-            assert ct in ('aggregated', 'per_asset', 'per_area'), ct
-        for at in area_types.values():
-            assert at in ('aggregated', 'per_asset'), at
-        self.cost_types = cost_types
-        self.area_types = area_types
-        self.units = units
-        self.deduct_abs = deduct_abs
-        self.limit_abs = limit_abs
-
-    def __call__(self, loss_type, values, area, number):
-        cost = values.get(loss_type)
-        if cost is None:
-            return numpy.nan
-        cost_type = self.cost_types[loss_type]
-        if cost_type == "aggregated":
-            return cost
-        if cost_type == "per_asset":
-            return cost * number
-        if cost_type == "per_area":
-            area_type = self.area_types[loss_type]
-            if area_type == "aggregated":
-                return cost * area
-            elif area_type == "per_asset":
-                return cost * area * number
-        # this should never happen
-        raise RuntimeError('Unable to compute cost')
-
-    def __toh5__(self):
-        loss_types = sorted(self.cost_types)
-        dt = numpy.dtype([('cost_type', hdf5.vstr),
-                          ('area_type', hdf5.vstr),
-                          ('unit', hdf5.vstr)])
-        array = numpy.zeros(len(loss_types), dt)
-        array['cost_type'] = [self.cost_types[lt] for lt in loss_types]
-        array['area_type'] = [self.area_types[lt] for lt in loss_types]
-        array['unit'] = [self.units[lt] for lt in loss_types]
-        attrs = dict(deduct_abs=self.deduct_abs, limit_abs=self.limit_abs,
-                     loss_types=hdf5.array_of_vstr(loss_types))
-        return array, attrs
-
-    def __fromh5__(self, array, attrs):
-        vars(self).update(attrs)
-        self.cost_types = dict(zip(self.loss_types, array['cost_type']))
-        self.area_types = dict(zip(self.loss_types, array['area_type']))
-        self.units = dict(zip(self.loss_types, array['unit']))
-
-    def __repr__(self):
-        return '<%s %s>' % (self.__class__.__name__, vars(self))
-
-costcalculator = CostCalculator(
-    cost_types=dict(structural='per_area'),
-    area_types=dict(structural='per_asset'),
-    units=dict(structural='EUR'))
-
-
-class Asset(object):
-    """
-    Describe an Asset as a collection of several values. A value can
-    represent a replacement cost (e.g. structural cost, business
-    interruption cost) or another quantity that can be considered for
-    a risk analysis (e.g. occupants).
-
-    Optionally, a Asset instance can hold also a collection of
-    deductible values and insured limits considered for insured losses
-    calculations.
-    """
-    def __init__(self,
-                 asset_id,
-                 taxonomy,
-                 number,
-                 location,
-                 values,
-                 area=1,
-                 deductibles=None,
-                 insurance_limits=None,
-                 retrofitteds=None,
-                 calc=costcalculator,
-                 ordinal=None):
-        """
-        :param asset_id:
-            an unique identifier of the assets within the given exposure
-        :param taxonomy:
-            asset taxonomy
-        :param number:
-            number of apartments of number of people in the given asset
-        :param location:
-            geographic location of the asset
-        :param dict values:
-            asset values keyed by loss types
-        :param dict deductible:
-            deductible values (expressed as a percentage relative to
-            the value of the asset) keyed by loss types
-        :param dict insurance_limits:
-            insured limits values (expressed as a percentage relative to
-            the value of the asset) keyed by loss types
-        :param dict retrofitteds:
-            asset retrofitting values keyed by loss types
-        :param calc:
-            cost calculator instance
-        :param ordinal:
-            asset collection ordinal
-        """
-        self.idx = asset_id
-        self.taxonomy = taxonomy
-        self.number = number
-        self.location = location
-        self.values = values
-        self.area = area
-        self.retrofitteds = retrofitteds
-        self.deductibles = deductibles
-        self.insurance_limits = insurance_limits
-        self.calc = calc
-        self.ordinal = ordinal
-        self._cost = {}  # cache for the costs
-
-    def value(self, loss_type, time_event=None):
-        """
-        :returns: the total asset value for `loss_type`
-        """
-        if loss_type == 'occupants':
-            return self.values['occupants_' + str(time_event)]
-        try:  # extract from the cache
-            val = self._cost[loss_type]
-        except KeyError:  # compute
-            val = self.calc(loss_type, self.values, self.area, self.number)
-            self._cost[loss_type] = val
-        return val
-
-    def deductible(self, loss_type):
-        """
-        :returns: the deductible fraction of the asset cost for `loss_type`
-        """
-        val = self.calc(loss_type, self.deductibles, self.area, self.number)
-        if self.calc.deduct_abs:  # convert to relative value
-            return val / self.calc(loss_type, self.values,
-                                   self.area, self.number)
-        else:
-            return val
-
-    def insurance_limit(self, loss_type):
-        """
-        :returns: the limit fraction of the asset cost for `loss_type`
-        """
-        val = self.calc(loss_type, self.insurance_limits, self.area,
-                        self.number)
-        if self.calc.limit_abs:  # convert to relative value
-            return val / self.calc(loss_type, self.values,
-                                   self.area, self.number)
-        else:
-            return val
-
-    def retrofitted(self, loss_type, time_event=None):
-        """
-        :returns: the asset retrofitted value for `loss_type`
-        """
-        if loss_type == 'occupants':
-            return self.values['occupants_' + str(time_event)]
-        return self.calc(loss_type, self.retrofitteds,
-                         self.area, self.number)
-
-    def __lt__(self, other):
-        return self.idx < other.idx
-
-    def __repr__(self):
-        return '<Asset %s>' % self.idx
-
-
-def get_values(loss_type, assets, time_event=None):
-    """
-    :returns:
-        a numpy array with the values for the given assets, depending on the
-        loss_type.
-    """
-    return numpy.array([a.value(loss_type, time_event) for a in assets])
 
 
 class RiskModel(object):
@@ -343,7 +144,7 @@ class Classical(RiskModel):
         n = len(assets)
         vf = self.risk_functions[loss_type]
         imls = self.hazard_imtls[vf.imt]
-        values = get_values(loss_type, assets)
+        values = asset.get_values(loss_type, assets)
         lrcurves = numpy.array(
             [scientific.classical(
                 vf, imls, hazard_curve, self.lrem_steps_per_interval)] * n)
@@ -488,7 +289,7 @@ class Scenario(RiskModel):
 
     def __call__(self, loss_type, assets, ground_motion_values, epsgetter):
         epsilons = [epsgetter(asset.ordinal, None) for asset in assets]
-        values = get_values(loss_type, assets, self.time_event)
+        values = asset.get_values(loss_type, assets, self.time_event)
         ok = ~numpy.isnan(values)
         if not ok.any():
             # there are no assets with a value
